Extract `ItemSelectorViewModel` to own file
